+++ app/src/main/java/com/crypto/katip/HomeActivity.java
@@ -1,18 +1,36 @@
 package com.crypto.katip;
 
-import android.content.Intent;
+
 import android.os.Bundle;
+import android.widget.TextView;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+import androidx.navigation.ui.AppBarConfiguration;
+import androidx.navigation.ui.NavigationUI;
 
 public class HomeActivity extends AppCompatActivity {
+
+    TextView notice;
+    BottomNavigationView navView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_welcome);
+        setContentView(R.layout.activity_main);
 
-        startActivity(new Intent(HomeActivity.this, LoginActivity.class));
+        notice = findViewById(R.id.notice);
+        navView = findViewById(R.id.nav_view);
+
+//        AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(R.id.navigation_profile, R.id.navigation_chat).build();
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
+//        NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
+        NavigationUI.setupWithNavController(navView, navController);
 
     }
+
 }
