+++ app/src/main/java/com/crypto/katip/LoginActivity.java
@@ -36,7 +36,6 @@ public class LoginActivity extends AppCompatActivity {
                     Toast.makeText(getApplicationContext(), loginResult.getError(), Toast.LENGTH_LONG).show();
                 } else if (loginResult.getSuccess() != null) {
                     Toast.makeText(getApplicationContext(), loginResult.getSuccess().getUsername(), Toast.LENGTH_LONG).show();
-                //    startActivity(new Intent(LoginActivity.this, HomeActivity.class));
 
                     finish();
                 }
