+++ app/src/main/java/com/crypto/katip/WelcomeActivity.java
@@ -1,0 +1,18 @@
+package com.crypto.katip;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class WelcomeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_welcome);
+
+        startActivity(new Intent(WelcomeActivity.this, LoginActivity.class));
+
+    }
+}
