+++ app/src/main/res/layout/activity_welcome.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".HomeActivity">
+    tools:context=".WelcomeActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
